Fifty-four upstream commits: the association read/write audit, the cluster
subsystem, the #683 cross-tenant entity isolation fix, and the read path made
provably write-free. Every fork customization is preserved; where upstream
fixed the same problem better, the fork's workaround was dropped in favour of
theirs rather than carried alongside it.

Migration numbering. Upstream's v5 (replication keyspace relocation, #726) and
v6 (vault-scoped entity records, #683) are registered here as v8 and v9. The
runner stores a single high-water mark and live data dirs are stamped at 6 by
this fork's own v5/v6, so renumbering the applied fork migrations instead would
have put upstream's below the mark, where they are skipped in silence. v7
(added in the prep commit) clears this fork's legacy 0x2F markers and must stay
below v8, which moves replication into that range. Upstream's two migration
tests are renumbered to match.

Superseded by upstream, workaround removed:
- entityVisibleInVault / entityPresentInVault and their four call sites. The
  fork gated the global 0x1F entity registry per vault in the engine; #683
  vault-scopes the records in storage, where the gate cannot be bypassed.

Kept, with upstream's improvement folded in:
- Embed counting. CountEmbedded/CountEmbeddableTotal now take upstream's #802
  vault parameter AND keep the fork's live-state filter, so EmbeddedCount can
  never exceed TotalCount at either scope. CountEmbeddedInVault gained the
  filter and a new CountLiveEngramsInVault is its denominator. The embed-status
  handler takes the vault param but keeps CountEmbeddableTotal, not Stat's
  EngramCount, as the denominator; upstream's test is realigned to that intent.
- Retroactive embed. Upstream's #644 vault-clear fence runs first and decides
  which engrams are still real; the fork's embedBisect then runs on what
  survives and decides which of those are actually bad content.
- Contradiction handling. The fork's durable 0x31 per-vault marker (the O(1)
  gate) and upstream's in-memory generation bump and debt readout are
  complementary and both fire, in batch.go and impl.go alike.
- Multi-vault MCP fan-out, now composing with upstream's UPSERT mode: a
  fan-out write suffixes the upsert key per vault, because the vaults get
  independent engrams and must not share one identity.
- Scoped API keys: multi-entry index deletion on revoke, over upstream's
  defer-Close refactor.
- gRPC jsonCodec, alongside upstream's chained not-Cortex interceptors.

RetryEmbedFailed now goes through refuseWrite rather than refuseAppend, so it
picks up the cluster single-writer guard upstream added.

Digest flags widen uint8 -> uint16 throughout, per upstream.

go build, go test and go test -race all clean.

+++ internal/engine/engine_reembed_test.go
@@ -36,7 +36,7 @@ func TestStartReembedVault_Success(t *testing.T) {
 	ctx := context.Background()
 
 	const vaultName = "reembed-success"
-	const DigestEmbed uint8 = 0x02
+	const DigestEmbed uint16 = 0x02
 
 	// Write a few engrams.
 	idStrings := make([]string, 3)
@@ -126,8 +126,8 @@ func TestRetryEmbedFailed_ClearsOnlyListedEngrams(t *testing.T) {
 	ctx := context.Background()
 
 	const vaultName = "retry-embed-failed"
-	const DigestEmbed uint8 = 0x02
-	const DigestEmbedFailed uint8 = 0x80
+	const DigestEmbed uint16 = 0x02
+	const DigestEmbedFailed uint16 = 0x80
 
 	idStrings := make([]string, 3)
 	for i := range idStrings {
